# Optimize object store interaction during job creation.