@jmchilton 2018-06-21
Doing more debugging and optimization and of course I came right back to this critical path. This makes a lot of sense for many outputs - but I think the benefit of this potentially skyrockets with large output collections. I wish I understood why I marked it as WIP, I'm going to close and reapply against dev and see what happens.